DDF-1698 Mark invalid metacards as invalid
Added a pre-ingest plugin to validate and mark metacards as invalid
or valid.
Added a pre-query plugin to check the query and if it is missing a
'valid-state' parameter, add `valid-state=true`.
Added a sample metacard validator to use for testing.
Edited metacard to allow setting metacards as invalid

DDF-1698 Refactor to use multivalued attributes

DDF-1698 Add itests

Added itests that use CSW to search for cards based on validation-warnings

DDF-1698 Only add warnings and errors on occurence

Previously validation-warnings and validation-errors were
created with the string "none". To find metacards with no warnings
or errors you would search for validation-warnings and
validation-errors = "none". Now you search for validation-errors
and warnings is null or not or specific validators

DDF-1698 Add federated integration test

Also some small fixups